[C] Move routes into helper function, remove asPath from NamedLink
