+++ components/global/AppHeader/index.tsx
@@ -1,27 +1,35 @@
-import React, { PropsWithChildren } from "react";
-import styled from "styled-components";
+import React from "react";
+import { useRouter } from "next/router";
+import { isActiveRoute } from "helpers/routes";
+
 import NamedLink from "components/atomic/NamedLink";
 import ServiceProviderBar from "components/global/ServiceProviderBar";
 import MainNav from "components/global/MainNav";
 import TopNavItem from "components/atomic/TopNavItem";
-
 import SignIn from "components/Auth/SignIn";
 import SignOut from "components/Auth/SignOut";
 
 function AppHeader({ children, className }: Props) {
+  const { asPath } = useRouter();
+
   return (
     <header className={className} role="banner">
       <ServiceProviderBar />
       <MainNav>
         <NamedLink route="communityList" passHref>
-          {({ active }) => <TopNavItem active={active}>Communities</TopNavItem>}
+          <TopNavItem active={isActiveRoute(asPath, "communityList")}>
+            Communities
+          </TopNavItem>
         </NamedLink>
-
         <NamedLink route="collectionList" passHref>
-          {({ active }) => <TopNavItem active={active}>Collections</TopNavItem>}
+          <TopNavItem active={isActiveRoute(asPath, "collectionList")}>
+            Collections
+          </TopNavItem>
         </NamedLink>
         <NamedLink route="itemList" passHref>
-          {({ active }) => <TopNavItem active={active}>Items</TopNavItem>}
+          <TopNavItem active={isActiveRoute(asPath, "itemList")}>
+            Items
+          </TopNavItem>
         </NamedLink>
         <span>
           <SignIn />
